Use PR strategy

+++ scripts/pr.sh
@@ -1,7 +1,9 @@
 #/bin/sh
 
-VERSION=$(echo $1 | sed -e 'sXrefs/heads/update-istio-XX')
+VERSION=$(echo $1 | sed -e 'sXupdate-istio-XX')
 URL=$2
+CURRENT_BRANCH=$1
+
 
 if [ -d $VERSION ]; then
     echo "Version $VERSION must be specified"
@@ -20,13 +22,14 @@ trap "cd $HERE && rm -rf $TMPDIR" EXIT
 
 ITEMS=$(ls -1 $HERE/charts/istio/$VERSION)
 
-git checkout istio-$VERSION
+git checkout $CURRENT_BRANCH
 git fetch origin 
-git switch -t gh-pages;true
-git switch gh-pages;true
-cp -f index.yaml $TMPDIR
-git checkout istio-$VERSION
+git switch gh-pages
+if [ -f index.yaml ]; then
+    cp -f index.yaml $TMPDIR
+fi
 
+git checkout $CURRENT_BRANCH
 git branch
 
 cd $TMPDIR
@@ -40,7 +43,11 @@ for ITEM in $ITEMS; do
     LAST_ITEM=/$ITEM-$ISTIO_VERSION.tgz/
     echo "Processing $ITEM $ISTIO_VERSION $CHART_URL"
     helm package $HERE/charts/istio/$VERSION/$ITEM
-    helm repo index  . --merge $TMPDIR/index.yaml --url $CHART_URL 
+    if [ -f $TMPDIR/index.yaml ]; then
+        helm repo index  . --merge $TMPDIR/index.yaml --url $CHART_URL 
+    else
+        helm repo index  . --url $CHART_URL 
+    fi
 done
 
 mkdir $HERE/build
@@ -54,7 +61,7 @@ cp -f $TMPDIR/index.yaml $HERE/pages
 cd $HERE
 
 if [ -n "GITHUB_OUTPUT" ]; then
-    echo "INDEX=$HERE/pages/index.yaml" >> "$GITHUB_OUTPUT"
+    echo "INDEX=$HERE/pages" >> "$GITHUB_OUTPUT"
     echo "VERSION=$VERSION" >> "$GITHUB_OUTPUT"
     echo "BUILDDIR=$HERE/build" >> "$GITHUB_OUTPUT"
 fi
